refac : #2 모달 컴포넌트 폴더로 이동
